Consolidate IdV cancellation feature specs
**Why**: To improve performance of feature specs, run 3 specs for IdV cancellation instead of 24.
